basic ndef message works
